refactor: allow mutiple renderer for geosearch
